requirements.txt
streamlit
pandas
altair
openpyxl
requests
ipywidgets
IPython.display 